add_entities_images: use internal call to lib/get_entities_images
